requirements.txt
Pillow
certifi
